Refactored code for article status list
